Check the machine name against a whitelist/blacklist
Instead of conflicting with Hægtesse and relying on device-specific
packages to bring in the right daemons, have Wys check the machine
name (/proc/device-tree/model) against a whitelist/blacklist pair to
determine whether it should run.  If it should not run, exit with a
status that systemd recognises as meaning it should not be restarted.